Update patch set 1
Patch Set 1:

> from reading the bug "https://bugs.eclipse.org/bugs/show_bug.cgi?id=486857"
 > and "https://git.eclipse.org/r/65510" now i understand that i'm not
 > alone with my thoughts about the GitProjectData and nested (nested
 > linked repos or submodules). The performance issues may accour in
 > my patches, because it's like big puzzle for me, no idea how it was
 > intended to work in EGit with multiple repositories or nested
 > ones...

Yup, that discussion made me look closer at what there is in GitProjectData, and then led to my trying to finish this. There's also some stuff in there that doesn't appear to be used at all (RepositoryChangeListener) that I plan to remove completely. Unless I can figure out what it was intended for and find a use for it, that is. Some of that code is very old...

An overall design document sure would help ;-)) But I guess I'd have to write that myself based on reverse-engineering, git history archaeology, and mind reading.

One thing I haven't figured out yet is how to do the whole synchronization stuff when there are submodules. But the synchronize view is only half-functional anyway :-(. (And I never use it.)

Patch-set: 1
Label: Verified=0