Update patch set 6
Patch Set 6:

> For package explorer & staging view. "> " doesn't show yet in the
 > repo view. Will have to check the label provider there.

The RepositoryViewLabelsProvider (also used in a number of dialogs/tables) just doesn't display the state of a file/resource, except on a top-level configured repository by prepending "> " if there's any kind of change in the repo. For submodules, it doesn't do so. Can be changed easily to also display the marker on submodule nodes. (But not right now -- too late.)

Displaying overlays/text decorations on file nodes in the working tree isn't done currently for normal repositories either. I'd consider that to be out of scope for this submodules topic; that would rather be a separate topic.

Patch-set: 6
Label: Verified=0