RepositoriesViewLabelProvider: mark dirty submodules
Prepend the dirty indicator "> " to the text label if the submodule
has changes.

Requires that the RepositoryCache ensures that all repositories are
registered under their normalized git directory file name (normalized
meaning not containing . or .. components). The
RepositoriesViewContentProvider uses a SubmoduleWalk, and that returns
non-normalized paths, while the RepositoryMappings always use normalized
ones. That results in two different repository instances for the same
git repository; listening for index diff changes on one won't trigger on
the other one, and thus labels in the repositories view wouldn't update.

Change-Id: Idf4002debdda94b35b278bff8194cde2ecba739c
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>