Safe preference reading of sash weights in GitHistoryPage
Preference values shall be treated as unverified external input and be
read in paranoid mode. Preferences can become corrupted; we can't trust
them. In no case shall garbled preferences cause the application to
fail.

Armor reading of sash weights and fall back to the default values if
the stored values cannot be read.

Bug: 488666
Change-Id: Ib22dbf855c9dde10108b6a90cc9ca7fcd7688966
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>