[WIP] Submodule support for connecting and disconnecting
RepositoryCache cleanup must make sure not to remove submodules and
nested modules prematurely. Since those are not registered as
"configured repositories", which would cause trouble when a repo was
deleted from the Repositories view anyway, we must make sure to remove
them only if no configured repository contains them.

This requires that the RepositoryCache keeps track of repository
nesting. To avoid duplicating code, the IndexDiffCache cleanup when a
configured repository is removed is now triggered from the
RepositoryCache.

Disconnecting a project: we must clear the session properties for the
RepositoryMappings. If the project is just disconnected, not closed, the
session properties would otherwise remain.

Connecting: first, if we re-connect a previously connected, then
disconnected project, the refresh during re-connection will not produce
a resource delta for all the project's contents (as that already exists
in the Eclipse resource tree). Thus the GitDataProvider would not re-add
RepositoryMappings for folders containing submodule or nested repository
working directories. Therefore touch all .git files and folders before
refreshing. And the SharingWizard had better do the refresh at all. The
GitLightweightDecorator must listen to mapping changes to trigger
re-decorations when new RepositoryMappings appear.

Activator: make auto-sharing also work if there are several repositories
above the project. Map the project to the closest, and enter the
farthest as configured repository.

AddCommand: map projects to the innermost repository found.

Change-Id: Idd0e34f70e9bc91fa9ab9e1e580bc483bcf2cdde
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>