Update patch set 2
Patch Set 2:

(In reply to comment #14)
> 
> Why did PS2 not trigger a Hudson build?
Hm, i don't know. Before i had always a build triggered automatically.
 
> I still don't quite understand the use case with Eclipse-linked git directories.
> Why do you want to do that?
We have a git extension that supports multi-repositories for git like it was before in ClearCase (similar to googles "repo"). It creates a "project" that is NOT a git repo with nested git repos inside:
s7p.project
+ s7p.eclipse
+ s7p.cpu1500
+ s7p.comm
...

s7p.eclipse is a git repo that holds the eclipse project only for faking eclipse to have a Git-Provider at top level (project).
s7p.cpu1500 is a real git repo that is linked in s7p.eclipse and must work with EGit without creating a eclipse project at top level to have EGit working.
s7p.comm same as s7p.cpu1500, there are actually 13 repos...

So the usage of the links was a workaround for our old version control in ClearCase and it worked ONLY that way. We did the same after migrating to Git, but here we had to create for each linked repo an eclipse project or organize it in working sets or the new "Project Explorer" views feature that shows them hiarchically instead of flat. All this was not accepted by the developers.

Now i'm trying to support this in EGit. But if the linked resources is a generall performance or handling issue. I may just create a empty repo at toplevel "s7p.project" without links and nested repo detection must work...

Do you think the isLinked() stuff will not be that easy to support?

Thank You!

Patch-set: 2
Label: Verified=0