Update patch set 2
Patch Set 2:

(1 comment)

> Do you think the isLinked() stuff will not be that easy to support?

I think this is opening a can of worms. The removal of the isLinked() check in GitResourceDeltaVisitor is a quite major change; may not be that easy to get all edge cases right. And, as you write yourself, all other isLinked() occurrences will need examination, and I fear we may still discover a lot of places that may just silently assume that resources are not linked. But perhaps I'm overestimating this... don't let yourself be discouraged by my reservations!

How are push & pull supposed to work with such linked repositories? If you do a commit/push/pull on the s7p.eclipse project, what shall be committed/pushed to which repository?

From your use case description, I still don't see why s7p.eclipse needs to link (and that's Eclipse linked resource, not symlink, right?) to s7p.cpu1500 and the rest. Off-hand, it looks to me that s7p.cpu1500 and the others could just be added as normal standard git submodules to the s7p.eclipse repo.

(Yes, I know that push & pull don't work yet for submodules in folders. That might be perhaps be a reason to do as you do currently. That's coming soon, though.)

BTW, I triggered a Hudson build manually.

Patch-set: 2
Label: Verified=0