Update patch set 3
Patch Set 3:

(In reply to comment #23)
> It looks like this adds another feature similar to but different from submodules
> which will _only_ work in Eclipse and seems to increase complexity. There is
> probably no way to use this from JGit or native git or do I miss something ?
The links are only related to Eclipse it self and we use them only for Eclipse. So this is not related to JGit or native Git and cannot be used there. This "feature" is used to have an EGit managed repo in project root to have EGit-Provider start looking at nested but linked repositories that are physically located in other locations. So, yes this is new for EGit, but already long there for Eclipse itself and is supported by other prlugins. As i said i can investigate the other solution with "real" nested but not linked repos or submodules now, because Thomas fixed it!

> I think we should rather focus on improving submodules support instead.
Yes, it's now more an more usable and i think nested repos (not submodules and not linked) should work too.

> If you need repo, this is available in jgit command line and we could consider
> to integrate it into EGit.
Thanks, but our company has an own tool and developers like me have to use them first :( but i want to investigate submodules in future.

Patch-set: 3