RepositoryCache: do not prematurely remove submodules
RepositoryCache cleanup must make sure not to remove submodules and
nested modules prematurely. Since those are not registered as
"configured repositories", which would cause trouble when a repo was
deleted from the Repositories view anyway, we must make sure to remove
them only if no configured repository contains them.

This requires that the RepositoryCache keeps track of repository
nesting. To avoid duplicating code, the IndexDiffCache cleanup when a
configured repository is removed is now triggered from the
RepositoryCache.

Change-Id: I3f437caccd58d6c9fb4187f66d9f53e7834a5224
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>