Update patch set 4
Patch Set 4: Code-Review-2

Not nice. That failure actually highlights a true bug. Looks like the attempt to fix bug 483664 in https://git.eclipse.org/r/#/c/62066/ was wrong.

There can actually be top-level repositories that are not in the configured repositories. Those get evicted from the cache too early (at any change in the "configured repositories" pref, which is wrong), but may be still referenced from some project's RepositoryMapping. Later something (didn't track it down exctly, probably the decorator) creates a new IndexDiffCacheEntry for this repository that isn't in the cache. Ergo that new IndexDiffCacheEntry is never removed. Moreover, in the SubmoduleFolderTest setup, we then look up such a repository again (childRepository, line 93 in this patch), and now we have a second RepositoryInstance for the same git directory, which gets its own IndexDiffCacheEntry (and which later is properly removed).

The test failure lists so many IndexDiff cache entries because they also don't get removed when the projects that reference them via their RepositoryMapping eventually get deleted. So cache entries accumulate over all the tests.

Andrey's approach in https://git.eclipse.org/r/#/c/62065/ is apparently better, but it's not quite correct yet either. The index diff cache entry should probably be removed only if removeProj == true, otherwise we might also be left with projects that still have RepositoryMappings to that repo. At least his approach wouldn't remove them from the cache, so as long it's referenced from somewhere, a lookup() would always return the same instance.

But in addition to his approach, we'll also have to remove index diff cache entries when the last project referencing a repo is closed or deleted.

So instead of all this complicated nesting detection, I think the correct way is to produce a change that
* reverts https://git.eclipse.org/r/#/c/62066/
* and removes index diff cache entries when repositories are regularly pruned() from the repository cache once they're no longer referenced.

The latter, however, will mean that the IndexDiffCache and its entries must not hold on to strong references to repositories either. Which means either more weak references, or key and store only git dirs, and get the repository where needed from a new RepositoryCache method getRepository(File gitDir), which will _not_ create the repo if not in the cache. Which also means that the index update jobs will have to be prepared to get null for the repo they're supposed to run on.

This might take a while to get right...

Patch-set: 4
Label: Code-Review=-2