Support copy/move of workspace if Git repository is under workspace
If git repositories are located under the workspace path moving or
copying the workspace broke the repository path information in
.metadata\.plugins\org.eclipse.core.runtime\.settings\org.eclipse.egit.core.prefs
broke since all repository paths were stored as absolute paths and paths
of repositories located under the moved or copied workspace weren't
matching this persisted paths anymore.

Fix this by storing the repository path relative to the workspace root
if the repository path is located under the workspace root. For
repositories not located under the workspace root still store the
absolute path since otherwise their path would break if the workspace is
moved or copied.

Bug: 358285
Change-Id: Ib73b76eb1d63587a767fec59c076fbe51c75e2f1
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>