Update patch set 3
Patch Set 3:

Do we have to care about backwards compatibility here? Once the pref is stored with relative paths and the user tries to open the workspace with an Eclipse that has an EGit that doesn't resolve upon loading, we'll end up with strange effects as File.getAbsolutePath() will resolve against the process's current working directory. Is that always the workspace root?

Patch-set: 3
Label: Verified=0