Update patch set 6
Patch Set 6:

> Hm. Maybe that test needs to be @Ignore'd. Weak reference
 > reclamation is not really testable...

Maybe it is... I guess the instability here is due to the core Activator's RepositoryChangeScanner. Just waiting for jobs in the test is not good enough; the scanner could kick in right after that, and then will hold a hard reference for some time. This test has to temporarily block that scanner. Time for a background job rule.

Patch-set: 6
Label: Verified=0