Update patch set 10
Patch Set 10:

> We already have org.eclipse.jgit.lib.RepositoryCache couldn't we
 > just use that ?

I had thought about that, too. The cache in JGit is a real cache, using soft references. It is designed to keep repository instances in the cache as long as possible (memory and the garbage collector permitting), even if no longer referenced from outside.

The EGit "cache" is not really a cache. It is a _registry_ of currently used repository instances, and it should forget instances once they're no longer referenced from outside.

Patch-set: 10