Fix potential NPE in IndexDiffCacheEntry
It's possible that an IndexDiffCacheEntry is disposed while an
IndexDiffUpdateJob is running. While we do cancel such a job, we do not
join on it. We do, however, null out the previous IndexDiffData. A still
running job therefore needs to be careful about using this property.

I've seen this occasionally in our Hudson logs, the last time in [1].

[1] https://hudson.eclipse.org/egit/job/egit.gerrit/8298/console

Change-Id: I88165cc20d4886444009c6654293ff461a777b02
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>