Update patch set 2
Patch Set 2:

A bit of background. I posted this change to illustrate which behavior would be needed. I our case, we use the FetchGerritChangePage in order to allow users to fetch a change. 
When an error occurs during the fetch operation, for instance if the user doesnt have permission, or the remote selected in the wizard screen is no valid, an error is reported but appears so quickly that the user cannot see it.
This is the idea behind this change.
If you have another method to suggest for resolving this issue, please provide it.
Thanks

Patch-set: 2
Reviewer: Gerrit User 1231 <1231@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0