Remove session properties when disconnecting a project
Disconnecting a project left the session properties for the
RepositoryMappings still attached to the resource info in the Eclipse
resource tree.

This prevents timely garbage collection of no longer used repository
instances because the mappings contain hard references to repositories.

Moreover, potentially stale properties might become active again when
the project is re-connected.

Change-Id: I845c708d3d98ba285c8db94edd2ed14deebd29fb
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>