Prepare Neon target platform
Use the Neon target platform as the default target platform and
explicitly set target platform for org.eclipse.egit.core.test.

Update org.junit to 4.12 since 4.11 is not available in the Neon orbit
version. Remove the unnecessary Maven dependency to junit declared in
pom.xml, it is already declared in manifests and available in the EGit
target platform.

Neon Eclipse platform requires BREE 8 so we have to use Java 8 for the
EGit build (for the compiler settings we still stick to source and
target 1.7 since we want to still support Java 7) otherwise unpacking
platform pack200 archives will fail since they have been built
using Java 8 and hence cannot be unpacked using Java 7's unpack200.

This requires to update the findbugs maven plugin to the latest version
which supports Java 8.

Ignore a couple of tests in ResourceUtilTest which now fail since
bug 476585 was fixed in Neon M6.

CQ: 10693
Change-Id: I013e27c7d96910907604bf5aa729cb651f64620a
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>