Prepare Neon target platform
Use the Neon target platform as the default target platform and
explicitly set target platform for org.eclipse.egit.core.test.

Since Neon Eclipse platform requires BREE 8 we have to also use
it for the EGit build (for the compiler settings we still stick
to source and target 1.7) since otherwise the tests will fail.

Update the findbugs plugin to the latest version which supports
Java 8.

Change-Id: I013e27c7d96910907604bf5aa729cb651f64620a
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>