When running "Commit" action open staging view instead of commit dialog
Add a preference to allow switching back to the old behavior of the
commit action to open the commit dialog. By default it now opens the
staging view. After opening the staging view auto-select and set focus
on unstaged files if there are any, otherwise set focus on the commit
message.

The commit dialog is still used by CleanupUncomittedChangesDialog.

Bug: 490121
Change-Id: I604f31268f83ab11c3f3a869edf3c7121f0e6478
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>