Prefer the usage of staging area / to stage over index or add to index
Matthias Sohn the EGit project leads suggested to me for my Git book to
used the term "Stagging Area" instead of Index. I think the terminology
in the Eclipse Git tooling should also be adjusted to reference that.

This is the first part of a series of changes. If accepted Iäm going to
replace the other reference to index with staging area.


Bug: 481360
Change-Id: Iad0f7e93dbaac1f684995f1fdb4546e3bc9c782f
Signed-off-by: Lars Vogel <Lars.Vogel@vogella.com>