Fix wrong size of 'commit and push' button
When switching between gerrit enabled and non gerrit enabled
repositories, the 'commit and push' button adds an icon, but does not
resize, therefore the label does not fit into the button width.

This could be fixed by always resizing the parent area containing the
buttons when switching between gerrit and non gerrit repositories, but
that would lead to the button width changing every time and the left
border of the buttons jumping around.

The only solution with correct size and without buttons visibly resizing
therefore seems to have an icon on the button both for the gerrit and
the non gerrit use case. I've taken the normal "Push" icon, since there
is no ready made icon with the meaning of both commit and push.

Bug: 491409
Change-Id: Ia5268c20478141646e9f17cb9ae44a054a1df5bf
Signed-off-by: Michael Keppler <michael.keppler@gmx.de>