Update patch set 1
Patch Set 1:

In bug 491464, the Gerrit refspec was added manually, so there's no guarantee it was "refs/for/". Might just as well (and just as valid!) be "refs/publish/". Shouldn't we at least handle that one, too? The third one "refs/drafts/" is probably unlikely to be used in a remote...

Patch-set: 1
Label: Verified=0