Allow deletion of branch hierarchy nodes
Up to now the user could not delete the branch hierarchy nodes. For
gerrit users this is a pain when trying to mass delete gerrit branches
with hierarchical layout being active.

The change expands the menu visibility, property testing and the
deletion command to explicitly handle the hierarchy branch nodes. A new
method has been introduced to get all refs of the hierarchy nodes
recursively. The deletion command then handles this list of refs,
instead of only taking one ref per selected node.

Overlapping selections (of branch nodes and their parent branch
hierarchy nodes) are handled correctly, because of the Set
implementation in the delete command.

Bug: 444869
Change-Id: I956954d07fcc44ca4a51ec3f681cf4f8ef97f86b
Signed-off-by: Michael Keppler <michael.keppler@gmx.de>