Fix wrong cursor position while editing URIs
When manually editing the URI of a remote, the automatic synchronization
of the URI and the remaining fields of the dialog (host, protocol, repo
etc.) leads to a wrong cursor position after each character being typed
by the user.

We avoid this by checking if the update of the URI is a no-op.

This fix applies for both the remote configuration dialog as well as the
clone wizard since they share the page with the implementation.

Bug: 492374
Change-Id: I39a20212782e7ef09f5e8df1a508e825b74817b8
Signed-off-by: Michael Keppler <michael.keppler@gmx.de>