Update patch set 3
Patch Set 3:

I'm not sure which dialog you mean, but the CheckoutDialog (RepositoryView context menu -> SwitchTo -> Other) displays and marks the currently checked out branch. So does the RepositoryView. I'm not sure if that is better from a usability perspective, than what you are suggesting, but I wanted to make Gitflow consistent with the rest of EGit.
With the same motivation, if there is another dialog that does not display the branch, maybe we should change it as well.

Patch-set: 3
Reviewer: Gerrit User 1638 <1638@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0