Exclude labels from tab stop
In the repository deletion dialog two labels are used to display the
paths of the repository and working tree. This change excludes them from
tab order, such that only the checkboxes and dialog buttons are focused
when using tab.

The labels can still be used for copying the path using the mouse.

Change-Id: Ib1a389d8040ca89cbdc490106aade65589bd31b3
Signed-off-by: Michael Keppler <michael.keppler@gmx.de>