Topic input on PushToGerritPage
Give the user a nicer way to define a topic when pushing to Gerrit.
Previously, a topic could be set only by appending "/<topic>" or
"%topic=<topic>" to the Gerrit branch.

Add a new text field for the topic, plus a new checkbox to select
whether a topic should be set at all.

State of both fields is persisted per local branch and restored when
the dialog is re-opened, unless the repository is in "detached HEAD"
state. This information is stored in the git config:

  [branch "someLocalBranch"]
    gerritTopicUse = [true, false]
    gerritTopic = <topic>

Topics must follow a few rules:
* they must not contain whitespace.
* they must not contain both commas and percent signs (either one is
  fine, though).
This is because of Gerrit restrictions. The page does validate the topic
field and prevents pushing if an invalid topic is specified (and "Use
topic" is selected).

If no topic is defined explicitly, it defaults to the local branch name.
Setting a topic explicitly overrides a possible topic specification
given in the branch field. If "Use topic" is selected and the topic
field is empty, the topic at Gerrit is cleared.

The topic field provides content assist listing the last 20 valid topics
used for the repository.

Bug: 449493
Change-Id: Idf22e0b8e4544e85cbfda5f1424a37eb746ade19
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>