[memory leak] make sure we only allow one GitHistoryPageSource instance
This should fix the issue that we can have multiple GitHistoryPage
instances for same repository created and held by GenericHistoryView.

The code in PageBookView.getPageRec(IWorkbenchPart) uses map with the
part as a key to get appropriate page instance. In case an existing page
is not found, the
GenericHistoryView.showHistoryPageFor(Object,
boolean, boolean, IHistoryPageSource) creates a new "dummy"
HistoryPageSourceWorkbenchPart, which was used as a key in the map
above. In the hashCode() and equals() methods the
HistoryPageSourceWorkbenchPart uses hashCode() and equals() methods from
IHistoryPageSource instances. Previously it could happen that multiple
GitHistoryPageSource instances were created by the GitAdapterFactory
because the factory instances are created for each adapter contribution
in the plugin.xml, so that GitHistoryPageSource instances returned from
different GitAdapterFactory instances were different.

With different GitHistoryPageSource instances the
HistoryPageSourceWorkbenchPart keys were always different (even for the
same repository/file), the "old" existing GitHistoryPage instances were
not found, new pages were created by GenericHistoryView and old pages
left in memory (until closing the history view).

Bug: 492099
Change-Id: I451c39589994387998053c5d09074ebf69cca0a4
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>