Leave the commit message enabled in staging view
The CommitActionHAndler tries to focus the staging view. If the view was
freshly created, all its widgets are disabled. Focus thus gets set to
nowhere, and context menus in the view where the commit command was
invoked get confused. ("Show In" and "Team" sub-menus in package
explorer, or the context menu in the repositories view.)

By leaving the commit message area always enabled, focus can be set
properly and all is fine.

Bug: 492839
Change-Id: Id082213c03f91b567990f7ecb983c85d1f8b0822
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>