Update patch set 3
Patch Set 3:

Tested this with a Mars.2 with EGit 4.3.1 including the old importer fragment without host bundle version constraint, and a maven-generated local repository.

Updating this Eclipse through "Install New Software..." from the local EGit update site gives the following:

If the Neon staging repo is not available:
* de-installs org.eclipse.egit.ui.importer
* upgrades EGit to 4.4.0
* does not install org.eclipse.egit.ui.smartimport (correctly, because this is >= 4.6)

(I don't quite understand why it de-installs importer if it then does not install smartimport...)

If the Neon staging repo is available:
* de-installs org.eclipse.egit.ui.importer
* Upgrades the Eclipse to Neon and EGit to 4.4.0
* installs org.eclipse.egit.ui.smartimport

I could not get upgrading via "Check for Updates" to work. Probably that does not connect to my local update site but to the one registered in the org.eclipse.egit-feature feature: http://download.eclipse.org/egit/updates . That one doesn't (yet) contain this p2.inf file and thus does not de-install org.eclipse.egit.ui.importer. If that's correct, we'd have to re-test that part once we do have this p2.inf in the official update site.

Patch-set: 3