Initialize dialog settings when a remote is configured
When a user cloned a repository via git:// and then configured the
remote for Gerrit through "Gerrit Configuration..." (specifying a
ssh:// push URI), the "Push to Gerrit..." wizard would still open with
the git:// URI pre-selected because the dialog hadn't remembered any
"last used" URI yet and "git://" is lexicographically smaller than
"ssh://". This was counter-intuitive.

Therefore: if a remote is configured for Gerrit, or a push or fetch
refspec is configured (and it's a Gerrit refspec), and the "Push to
Gerrit" or "Fetch Change from Gerrit" dialogs have no last used URI
set, initialize it to the URI just configured. This gives a reasonable
initial value when those dialogs are subsequently opened the first time
for that repository.

Bug: 493935
Change-Id: I9815358a5d2e0a31ce355513d5ad76139e6e123d
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>