Unset session property when a project is disconnected
To avoid excessive accesses to the persistent property for the
repository provider, ResourceUtil maintains its own cached information
about providers associated with projects in a session property. Ensure
that this property is cleared when a provider is unmapped (any provider,
might also be a non-git provider) to prevent that we keep stale
references.

Change-Id: I600ab2b8c38348c1820837eea1441cd6ffcbd889
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>