Test stability: wait for specific child node after expansion
The first test in GlobalConfigurationPageTest frequently fails on Hudson
because it doesn't find the child node "Configuration". It appears that
child nodes in the preference page's tree are added one after another,
so waiting until there is any child is not good enough: "Configuration"
is the second child and might not have appeared yet.

Change-Id: Ic5690b481b111fe98aa6d8da1108d5ee462c6e1a
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>