Paste in repo view need not log error
The PasteCommand shows an alert to the user _and_ logs an error. That's
overkill here. Just showing an alert is good enough. Besides, it's not
really an error.

Even better would be to deactivate the handler when the clipboard
doesn't contain a git repo location. But I suspect a property tester for
this would be rather expensive, so I opted for this simpler solution.

Also add a missing clipboard.dispose() in the FetchGerritChangePage.

Bug: 495629
Change-Id: I8d0dbf14ad168e38fe722827a67fcd58ef9f4cf8
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>