Fix NoWorkTreeException in GitImportWizard
When a bare repository is selected in the GitSelectRepositoryPage, the
import wizard throws and reports multiple NoWorkTreeExceptions.

Since a bare repository has no working tree, one cannot import from a
bare repository. Therefore, do not display bare repositories in the
GitSelectRepositoryPage at all when used in the GitImportWizard. The
user thus cannot select a bare repository in the first place.

Bug: 495930
Change-Id: Ib10d5f7305f8ebfb12545ae780847d9b78ce0295
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>