Refactor BranchOperationUI
This class packed too many things into one, resulting in hard to
understand flow of control. Also, a class with different "modes" in
anathema to object-orientedness and to separation of concerns.

Eliminate the different modes from BranchOperationUI: they all just
defined which dialog to open initially, if any. Create those dialogs
instead in the few places when the static factory methods were used,
then use the BranchOperationUI to do the checkout, if appropriate.

Simplify RenameBranchDialog by re-using the BranchRenameDialog
:-) and by not checking for tags, which are not displayed anyway.

Make both the asynchronous (start()) and the synchronous (run())
checkout operations perform the the same actions, including project
restoration.

Factor out the checking for running launches so that it can be re-used
elsewhere.

Change-Id: If83e7960253d99b04e4a476329647bd009ec8898
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>