Fix commit amend to show full commit when clicking "amend" option
In the Git Staging view, when editing an existing commit, it should be
possible to re-edit all details of a commit. This includes not only the
commit message, but also all file additions, modifications and
deletions. Therefore, the "Staged Changes" file list should list the
changes that happened in the commit that is being edited.

Bug: 353816
Bug: 459152
Bug: 490126
Change-Id: Idcd1efeeee8b3065bae36e285bfc0af24ab1e88f
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>