Update patch set 1
Patch Set 1:

I missed to add the change in build-help.xml referencing the 4.4 new and noteworthy. Bug https://bugs.eclipse.org/bugs/show_bug.cgi?id=496089. Apparently I had changed this locally when generating the documentation update contained in this change (it's replacing 4.3 new and noteworthy by the one from 4.4) but missed to add this modification to this commit.

Patch-set: 1