Update patch set 3
Patch Set 3:

(4 comments)

Looks good! I tested the functionality and can confirm it works well. Thanks!

I only have a few suggestions for UI:

* "Override preference with a specific merge strategy"

I'd rather put: "Use specific merge strategy..."

Most people will probably not be aware of the fact that there is a preference, so they'd ask themselves, which preference is overwritten.

I'd put ... behind the label, to indicate that this will open more settings.

* "Short Name"

Why not just use "Name" instead? "Short Name" only makes sense if there is a normal name as well.

* Provided by

I'm not sure if it is a good idea to put that in this dialog. Most people will not be interested in that information and it uses a lot of space. I think it would rather make sense to put this information in the preference page only.

* It would be great, if we could resize the dialog, if the checkbox is enabled to make more room for the table, especially with the Merge... dialog.

* Maybe this is just on Linux with GTK2, but the scrollbar of the table is partially hidden.

Patch-set: 3
Reviewer: Gerrit User 7214 <7214@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0