Update patch set 3
Patch Set 3:

(4 comments)

Thank you for your review Philip!
I fixed most of your UI comments. However, I don't want to spend time on layouting the table because SWT makes that ridiculously difficult. I'd have to introduce a lot of code for very little added-value I think and I don't want anybody would want to maintain this. So I leave it like that for now and wait for other reviewers feedback to see if I really must do something.
I didn't want to use a table in the 1st place because then the text would not wrap and also it's not possible to set help context id on table items.

Patch-set: 3
Label: Verified=0