GitHistoryPage: move the search widget into the toolbar
This necessitates being more careful with global action registrations.
These previously used updateActionBars(), which updates also the
toolbars, which messes up when the focus changes from the history page
to the search bar's text input: one needed two clicks to set the focus
on the search bar. Resolved by updating only the global menu manager
when we changed only global action handlers.

ToolbarManager has a number of problems correctly re-displaying a
changing toolbar with dynamic items. In particular, it is not sufficient
to only update the ToolbarManager. It is in this case necessary to use
updateActionBars() indeed, which does much more and manages to re-layout
and re-position the toolbar properly. However, it messes up with the
background of the Composite containing the ToolBar. This was worked
around by some ugly back-and-forth with the focus. Some more focus
juggling had to be done to get the history view (and thus its status
line manager) to activate when the focus was elsewhere and the user then
clicked directly into the search bar, thus focusing it.

The search result indicator ("1/n") had to be moved out of the toolbar
itself. Instead there's now a listener interface that gets handed
that text and that then can display it as it sees fit. In this
implementation, it's simply written to the status bar. I also had
experimented with using a tooltip, but that becomes a hassle very
quickly. And tooltips belong to clan MacLeod: there can be only one.

Bug: 496278
Change-Id: I5adc521d19b0b99ee98bc573a807c72ba7d5caed
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>