Update patch set 3
Patch Set 3:

With this change I can select a bare repository and try to add it from GitImportWizard and after adding it it's not shown in the list of repositories of the GitImportWizard.

I think we should either also hide bare repositories in the dialog used to add repositories to the list of repositories shown in GitImportWizard. 

Or we allow to import bare repositories and show a hint when a bare repository is selected that projects can't be imported from a bare repository. (We also allow to add bare repositories to the repositories view ...)

Patch-set: 3
Label: Verified=0