Remove extra line for the repository information in the Git staging view
The current placement of the repository information takes to much space.
During normal work, the developer knows which repo he is working on and
the place for the repository becomes wasted space.
This information fits well into the first section, as this reads as a
header for the whole UI.

Bug: 497197
Change-Id: I82977339f7045b8554e881880f20f3054b1f89da
Signed-off-by: Lars Vogel <Lars.Vogel@vogella.com>